[AArch64] Allow shrink-wrapping of non-leaf vector PCS functions
With the function ABI stuff, we can now support shrink-wrapping of
non-leaf vector PCS functions.  This is particularly useful if the
vector PCS function calls an ordinary function on an error path,
since we can then keep the extra saves and restores specific to
that path too.

2019-09-30  Richard Sandiford  <richard.sandiford@arm.com>

gcc/
	* config/aarch64/aarch64-protos.h (aarch64_use_simple_return_insn_p):
	Delete.
	* config/aarch64/aarch64.c (aarch64_components_for_bb): Check
	whether the block calls a function that clobbers more registers
	than the current function is allowed to.
	(aarch64_use_simple_return_insn_p): Delete.
	* config/aarch64/aarch64.md (simple_return): Remove condition.

gcc/testsuite/
	* gcc.target/aarch64/torture/simd-abi-9.c: New test.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@276340 138bc75d-0d04-0410-961f-82ee72b054a4